add token calculation for standalone openai text node
